fixed some errors with types

+++ vector3d/vector.py
@@ -1,8 +1,8 @@
 class Vector:
 	from math import sqrt
-	x = double()
-	y = double()
-	z = double()
+	x = float()
+	y = float()
+	z = float()
 	
 	def __init__(self, x=0, y=0, z=0):
 		self.x = x
